Delegate the loading of the web page to a method.
Instead of loading the web page in the main crawler method it is delegated to
a special method: _get_page.  In this way subclasses can overwrite this new
method in order to catch some special exceptions that might occur.  Currently
the Mangareader class is the only one making use of that.