Implement simple resume logic
By removing one url from the unpickled state we can know that the page
will be loaded again and can hope that we will thereby find updates on
the crawled page (new pages to crawl).

This only works as long as the user supplies an url for the same comic
on the command line as was previously downloaded and stored in the
pickle file.