Move all dump file handling to the Site class
The queue now has more "symmetric" functions with get_state() and
__init__().  The Site class did already have this kind of "symmetric"
functions with load() and dump() but now the actual work is done in
Site.dump() instead of Queue.dump().

The full Queue instance can not yet be dumped by the Site class as
Site.load() needs to be able to modify the state before recreating the
queue.