Remove type-ignore comment
Python allows to declare only the annotation for a name without a value.
The search for a suitable crawler class can then not check if the DOMAIN
is None but must check if we reached the base class which does not have
a value for DOMAIN.  Alternatively one could have used hasattr().