Set a minimum height for lazy images
Otherwise the <img> tags would have height 0 before the image is loaded.
This would result in all image tags being "close" to the last loaded
image.  This in turn results in all images being loaded at once as soon
as the user scrolls near any of the unloaded images.

By giving the images a minimum height they are not all close to each
other any longer.  This ensures that they can be lazy loaded in batches.